requirements.txt
python-dotenv
azure-identity
azure-ai-ml
azure-ai-projects
openai


